Restructure game source files into libraries
