+++ lib/game/board.h
@@ -8,7 +8,7 @@
 #include <memory>
 
 #include "lib/grid.h"
-#include "src/tile.h"
+#include "tile.h"
 
 namespace pie {
